перенес методы для тестирования в модель игры
